Proposal changed due the #2 revision
The primary issue with the current proposal is that it's not at all clear what you'll be predicting. 
Response: My proposal is to predict the "overall" feature.


The labels haven't been created yet, so you'll need to have labels for the project.
Response: What i called label, is the "overall" feature, that indeed exists in the dataset but for my strategy is to remove this data and run the model to predict this feature.
After that it will be compared with the removed feature in order to test the efectiveness of the model to predict this information. In the section "Solution Statment" i mentioned that.

 You'll also need to be able to  pick specific metrics that you'll use to quantify the results of the supervised learning models. If the labels are continuous,
 then you'll need to pick one or more appropriate regression metrics. If the labels are discrete, then you'll need to choose appropriate classification metrics.
Response: The metric will be how the predicton of the feature "overall" will be closer of the original information.